requirements.txt
pathway[all]
python-dotenv==1.0.1
litellm>=1.35
Google-generativeai
Sentence-transformers